Use custom save method instead of auto_now and auto_add_now parameters
auto_add and auto_add_now parameters can be evil.
For instance if we have two fields, like in this case, with auto date time
the two date times will NOT be identical on save. This is due to both
fields calling timezone.now() internally just nanoseconds after each other
and not at the same time. If the date times are then compared they will differ.
This can be a bit confusing when having one auto_now_add and another auto_now
field like in our case. The created and modified fields would NOT have the same
time when the Note is created.

This has been reported long ago to the Django dev team:
https://code.djangoproject.com/ticket/16745

And even the core team does not seem to like the auto date fields:
http://comments.gmane.org/gmane.comp.python.django.scm/128505